requirements.txt
asgiref==3.7.2
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
distlib==0.3.8
Django==5.0.1
django-ratelimit==4.1.0
djangorestframework==3.15.1
executing==2.0.1
filelock==3.13.1
ipython==8.19.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
ratelimit==2.2.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
virtualenv==20.25.0
wcwidth==0.2.12
